+++ Gemfile
@@ -1,0 +1,7 @@
+source 'https://rubygems.org'
+
+group :development, :test do
+  gem 'rake'
+  gem 'pry'
+  gem 'rspec'
+end
