+++ Rakefile
@@ -1,0 +1,3 @@
+$:.push(File.dirname(__FILE__))
+
+import *Dir.glob('tasks/*.rake')
