+++ dynamic_binding.gemspec
@@ -1,11 +1,11 @@
 Gem::Specification.new do |s|
   s.name        = 'dynamic_binding'
-  s.version     = '0.1'
+  s.version     = '0.2'
   s.date        = '2012-04-23'
   s.summary     = "Rebind Procs for fun and profit."
   s.description = "A small Gem implementing a certain form of dynamic binding"
   s.authors     = ["Niklas Baumstark"]
   s.email       = 'niklas.baumstark@gmail.com'
-  s.files       = ["lib/dynamic_binding.rb"]
+  s.files       = ["lib/**/*.rb"]
   s.homepage    = 'https://github.com/niklasb/ruby-dynamic-binding'
 end
