+++ spec/spec_helper.rb
@@ -1,0 +1,2 @@
+require 'rspec'
+$:.push(File.join(File.dirname(__FILE__), '..'))
